@phated 2020-09-22
I dig this. I think it would be great to have the benchmarks and a test for the fix you added.